moved tests from decode_test to slicedecoder_test
